Fixes #182
No longer necessary to use a secondary method to save a sensor readings in batch, by DDD and as Readings are a value object of the Sensor' aggregate, in order to save each Reading we can call save(Sensor) to persist each Reading, as now the Readings and its Respective list and persisted embedded.

It is no longer needed to manually set the Reading' sensor as now the connection between the two is persisted. Given this attribute Sensor in the Reading Class became unused and for that it was removed.

ReadingRepository deleted as we no longer need a Readings repository, as it as an Value Object does not require a repository, as the root repository from the aggregate is being used.

The ReadingList class is now being persisted embedded through it's sensor's, and so are each Reading's creating a secondary table in the DB where the relationship between the sensor and it's readings can be seen, through a Sensor List of Readings Table.

This is achieve using @ElementCollection annotation to persist the List<Reading> as a One to Many relationship.