Addresses #183
Most simple and direct implementation Java' System.currentTimeMillis( ) existing method which returns the current time in milliseconds as a long;
The only required math is a subtraction between the initial and final values;

Question: When/If the milliseconds counter goes to zero what happens in the application timer? Whats the right solution to apply?